% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_accuracy}
\alias{dummy_accuracy}
\title{Generate (uncorrelated) dummy data (target: accuracy)}
\usage{
dummy_accuracy(n = 100, theta = seq(0.7, 0.9, 0.05))
}
\arguments{
\item{n}{integer, sample size (length 2: diseased healthy)}

\item{theta}{numeric, accuracy = theta}
}
\value{
dummy data (target: accuracy)
}
\description{
Generate (uncorrelated) dummy data (target: accuracy)
}
