% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_sensspec}
\alias{dummy_sensspec}
\title{Generate (uncorrelated) dummy data (target: sensspec)}
\usage{
dummy_sensspec(n = c(50, 100), theta = seq(0.7, 0.9, 0.05))
}
\arguments{
\item{n}{integer, sample size (length 2: diseased healthy)}

\item{theta}{numeric, sens = theta; spec = rev(theta)}
}
\value{
dummy data (target: sensspec)
}
\description{
Generate (uncorrelated) dummy data (target: sensspec)
}
