% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate model performances and covariance matrix.}
\usage{
estimate(comparison, method = "default", ..., args = NULL)
}
\arguments{
\item{comparison}{SEPM.comparison, see \code{\link{compare}}.}

\item{method}{cahracter, speciying known estimation method (default: "default"). Alternativly a
custom SEPM.control object may be spcified, see \code{\link{estimate_control}}.}

\item{...}{further (named) arguments to control$mapping.}

\item{args}{named list of further arguments to control$mapping}
}
\value{
An SEPM.comparison object which summarizes the estimates result.
}
\description{
This function allows the estimation of model performances and their covariance by employing a
prespecified estimation method.
}
\details{
When no arguments are supplied other than comparison, maximum likelihood
(control = "default") is conducted. Call \code{\link{estimate_control}}() for more information regarding
estimation methods.
}
\examples{
comp <- compare("accuracy", comparison=matrix(rbinom(30, 1, 0.8), 10))
estimate(comp)
}
