% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_control.R
\name{estimate_control}
\alias{estimate_control}
\title{Statistical Evaluation of Prediction Models (SEPM)}
\usage{
estimate_control(method, type = c("generic", "probability"), mapping)
}
\arguments{
\item{method}{character, name of the estimation method}

\item{type}{either \code{"generic"} (default) or \code{"probability"}, the latter should only be
used for classification tasks}

\item{mapping}{a function which specifies how the estimation is conducted, see details.}
}
\value{
A SEPM.estimate.control object which is passed to \code{\link{estimate}}.
}
\description{
Control the estimation process.
}
\details{
If \code{type="generic"}, mapping needs to be a function with input \code{comparison}
and output \code{list(theta.hat = ..., sigma.hat = ...)} which is computed based on the input
comparison matrix. If \code{type="probability"}, mapping needs to to a function with input
argument \code{p} (a probaility (matrix)) and output a modified (e.g. shrunk) probability
with the same dimensions as \code{p}.
}
