% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_feature}
\alias{get_feature}
\title{Extract features from a list}
\usage{
get_feature(x, what = "model.name")
}
\arguments{
\item{x}{list or list-like object}

\item{what}{slot to extract from x}
}
\value{

}
\description{
Extract features from a list
}
