% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\name{guide}
\alias{guide}
\title{Guidance for model evaluation study.}
\usage{
guide(x)
}
\arguments{
\item{x}{an arbitrary R object.}
}
\value{
No return value, only a message is printed to guide the user.
}
\description{
Use guide(x) or x \%>\% guide() for any SEPM object x (SEPM.hypothesis,
SEPM.estimate, ...) to obtain hints which function to call next for your model evaluation study.
Get started by calling guide().
}
\examples{
guide()
}
