% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Statistical Evaluation of Prediction Models (SEPM)}
\usage{
report(evaluation, ...)
}
\arguments{
\item{evaluation}{SEPM.evaluation object}
}
\value{
NULL, see details
}
\description{
Report the results of the evaluation study with R markdown.
}
\details{
... Create R Markdown template
}
\examples{

}
