% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Statistical Evaluation of Prediction Models (SEPM)}
\usage{
select(
  comparison,
  method = c("best", "close", "optimal", "user"),
  ...,
  args = list()
)
}
\arguments{
\item{comparison}{SEPM.comparison}

\item{method}{character, specifying how models (algorithms) for the evaluation study shall be
selected. See}

\item{...}{further named arguments to selection method}

\item{args}{list of named arguments to selection method}
}
\value{
indices of selected model
}
\description{
Select models based on hold-out-validation during the learning phase.
}
\details{
Select models based on empirical performances prior to evaluation study, see
\code{\link{select_best}}, \code{\link{select_close}}, \code{\link{select_optimal}}
}
