% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_best}
\alias{select_best}
\title{Return empirically best prediction model}
\usage{
select_best(
  comparison,
  mode = c("weighted", "min"),
  threshold = c(0.5, 0.5),
  regu = 1,
  break_ties = TRUE,
  ...
)
}
\arguments{
\item{comparison}{SEPM.comparison}

\item{mode}{character, "weighted" or "min"}

\item{threshold}{numeric, length corresponds to number of subgroups}

\item{regu}{numeric, regularization parameter}

\item{break_ties}{logical, default: TRUE}

\item{...}{further arguments (currently ignored)}
}
\value{
a vector of selected model indices
}
\description{
Return empirically best prediction model
}
