% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_close}
\alias{select_close}
\title{Select prediction models close to the best}
\usage{
select_close(
  comparison,
  k = 1,
  mode = c("weighted", "min"),
  threshold = c(0.5, 0.5),
  max_models = Inf,
  regu = 1,
  break_ties = TRUE,
  ...
)
}
\arguments{
\item{comparison}{SEPM.comparison}

\item{k}{numeric, multiplier}

\item{mode}{character, "weighted" or "min"}

\item{threshold}{numeric, length corresponds to number of subgroups}

\item{max_models}{integer, maximum number of models to be selected}

\item{regu}{numeric, regularization parameter}

\item{break_ties}{logical, default: TRUE}

\item{...}{further arguments (currently ignored)}
}
\value{
a vector of selected model indices
}
\description{
Select prediction models close to the best
}
