% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_control.R
\name{select_control}
\alias{select_control}
\title{Statistical Evaluation of Prediction Models (SEPM)}
\usage{
select_control(method, mapping)
}
\arguments{
\item{method}{character, name of the selection method}

\item{mapping}{a function with formal input arguments \code{estimate} (vector of parameter
estimates), \code{cov} (covariance matrix of \code{estimate}) and \code{h}
(an \code{SEPM.hypothesis}. The output of mapping needs to be a logical with same length as
\code{estimate} where \code{TRUE} represents a selected and \code{FALSE} a non-selected model.}
}
\value{
A SEPM.select.control object which may be passed to \code{\link{select}}.
}
\description{
Control the selection process.
}
\details{
mapping can have further input arguments which need to be specified via ... or args when
calling \code{\link{select}}. A call of \code{select_control()} prints all known selection methods.
}
\examples{
select_control()
select_control("rank")

}
