% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_optimal}
\alias{select_optimal}
\title{Select prediction models in an optimal fashion for the evaluation study}
\usage{
select_optimal(
  comparison,
  method_select = "few",
  method_order = "param",
  target_order = "prob",
  combine_order = "min",
  max_models = Inf,
  n_eval = 100,
  prev_eval = 0.5,
  n_val = NA,
  prev_val = NA,
  rdm = TRUE,
  threshold = c(0.75, 0.75),
  batch_size = 10,
  max_iter = 50,
  target_tol = 1e-04,
  method_ext = "basic",
  method_pred = "mbeta_approx",
  return_simvals = FALSE,
  steady_plot = FALSE,
  save_plot = FALSE,
  info_plot = FALSE,
  ylim_plot = NULL,
  ...
)
}
\arguments{
\item{comparison}{SEPM.comparison}

\item{method_select}{"few" or "max"}

\item{method_order}{"param" or "sample"}

\item{max_models}{integer, maximum number of models}

\item{n_eval}{evaluation sample size}

\item{prev_eval}{evaluation prevalence}

\item{n_val}{validation sample size}

\item{prev_val}{validation prevalence}

\item{rdm}{logical, simple random sampling (TRUE) or case-control sampling (FALSE)}

\item{threshold}{numeric, length identical to number of groups}

\item{batch_size}{integer}

\item{max_iter}{integer}

\item{target_tol}{numeric, tolerance parameter}

\item{method_ext}{"basic" or "cov"}

\item{method_pred}{"mbeta_approx" or "rmvbin"}

\item{return_simvals}{logical}

\item{steady_plot}{logical}

\item{save_plot}{logical}

\item{info_plot}{logical}

\item{ylim_plot}{numeric, length 2}

\item{...}{}

\item{param_order}{"mean", "tstat" or "prob"}
}
\value{
a vector of selected model indices
}
\description{
Select prediction models in an optimal fashion for the evaluation study
}
