% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEPM.R
\docType{package}
\name{SEPM}
\alias{SEPM}
\title{Statistical Evaluation of Prediction Models (SEPM)}
\description{
Estimate and evaluate the performance of (machine learned) prediction models
}
\section{Purpose}{


Model evaluation... Divided into comparison, estimation, inference
}

\section{Model evaluation}{


The main ...
}

\section{Model selection}{


A second ...
}

