% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer}
\alias{infer}
\title{Hypothesis testing regarding model performances}
\usage{
infer(estimation, method = "maxT", transform = "none", lfc = "t")
}
\arguments{
\item{estimation}{object of class SEPM.estimation (see \code{\link{estimate}}).}

\item{method}{character; specify the (multiple) test used (default: "maxT", currently the only choice)}

\item{transform}{character; specify a known transform, e.g. "none" (default), "logit" or
"asin.sqrt". Alternativly a object of class SEPM.transform created via
\code{\link{define_transform}}).}

\item{lfc}{}
}
\value{
An SEPM.inference object, including inference results, parameters estimates and
hypothesis.
}
\description{
This functions allows hypothesis testing regarding a hypothesis system based on some previously
derived parameter estimates (see \code{\link{estimate}}).
}
\examples{
set.seed(1)
y <- rep(0:1, each=100)
yhat <- sapply(seq(0.7, 0.9, 0.05),
 function(p){ifelse(rbinom(200, 1, p)==1, y, 1-y)})
define.hypothesis("accuracy", 0.8) \%>\%
 compare(predictions=yhat, labels=y) \%>\%
 estimate() \%>\%
 infer() \%>\%
 summary()
}
