% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEPM.R
\docType{package}
\name{SEPM}
\alias{SEPM}
\alias{SEPM-package}
\title{Statistical Evaluation of Prediction Models (SEPM)}
\description{
The SEPM package allows to conduct statistical inference regarding the the performance of (machine learned) prediction models.
It has two main functionalities, which are briefly described below. Note that this is an early development version.
}
\section{Model selection}{


Select a (sub)set of models from the initially large set of initial candidate models.
This is usually done based on some prior information, in particular hold-out- or cross-validation estimates.
}

\section{Model evaluation}{


Given a collection of selected models, the goal is now to evaluate them on independent test data.
Besides performance estimation, simulataneous confidence regions and test decisions may also be obtained.
}

